Make type changes for solo code explicit in test/minigzip.c.

+++ test/minigzip.c
@@ -252,7 +252,7 @@ static int gzwrite(gzFile gz, const void *buf, unsigned len) {
         (void)deflate(strm, Z_NO_FLUSH);
         fwrite(out, 1, BUFLEN - strm->avail_out, gz->file);
     } while (strm->avail_out == 0);
-    return len;
+    return (int)len;
 }
 
 static int gzread(gzFile gz, void *buf, unsigned len) {
@@ -269,7 +269,7 @@ static int gzread(gzFile gz, void *buf, unsigned len) {
     strm->next_out = (void *)buf;
     strm->avail_out = len;
     do {
-        got = fread(in, 1, 1, gz->file);
+        got = (unsigned)fread(in, 1, 1, gz->file);
         if (got == 0)
             break;
         strm->next_in = in;
@@ -283,7 +283,7 @@ static int gzread(gzFile gz, void *buf, unsigned len) {
         if (ret == Z_STREAM_END)
             inflateReset(strm);
     } while (strm->avail_out);
-    return len - strm->avail_out;
+    return (int)(len - strm->avail_out);
 }
 
 static int gzclose(gzFile gz) {
